fix: correct migration patch path and clean up orphaned Module Def
