+++ block_administrator/patches.txt
@@ -4,4 +4,4 @@
 
 [post_model_sync]
 # Patches added in this section will be executed after doctypes are migrated
-block_administrator.patches.v1.0.0.add_block_administrator_login_field
+block_administrator.patches.v1_0_0.add_block_administrator_login_field
